ok nu alleen nog de comments verder is ie perfect.
